possibily to erase a file

+++ backend/controllers/book.controllers.js
@@ -1,4 +1,5 @@
 const Book = require('../models/book')
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book);
@@ -21,11 +22,24 @@ exports.createBook = (req, res, next) => {
       .catch(error => res.status(400).json({ error }));
   };
 
-  exports.deleteBook = (req, res, next) => { //supprimer un livre
-    Book.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json ({ message: 'Livre supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
-  };
+  exports.deleteBook = (req, res, next) => { //supprime un livre
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
   exports.getOneBook =(req, res, next) => { //trouver un livre
     Book.findOne({ _id: req.params.id })
